refactor: Move <PDFElement/> to @chainlit/components
